build: Simplify setting local value of flycheck-emacs-lisp-load-path
This variable is set as a risky local variable by `flycheck`. However it's
beneficial to have it set to a different value for this project. Use the
regular setting, such that Emacs will always warn users when the variable is
set to a local value. To avoid the latter warning, one can add something like:
```emacs-lisp
(put 'flycheck-emacs-lisp-load-path 'risky-local-variable nil)
```
to their Emacs config.